define black in Data.Colour
The reason for adding black is that (foo `over` mempty) is sort of ambigous.
Granted, the haskell type inference can often figure out this type and
decide if mempty should be Colour or AlphaColour; however, it is much less
clear for a human reader.  And given that (foo `over` black) and
(foo `over` transparent) have signifigantly different results, it is useful
for the user to be able to easily specify which one is meant.

The transparent value already exists.  Here we are adding the black value.